compose/requirements.txt
Django==3.1.7
django-celery-beat==1.4.0
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-spectacular==0.13.2
gunicorn==20.0.4
isort==5.7.0
mysqlclient==2.0.3
opentracing==2.4.0
pre_commit==2.10.1
pytest==6.2.2
pytest-cov==2.11.1
pytest-django==4.1.0
pytz==2021.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
whitenoise==5.2.0
